requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
constantly==23.10.4
cookiecutter==2.5.0
cryptography==42.0.2
cssmin==0.2.0
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-appconf==1.0.6
django-browser-reload==1.12.1
django-ckeditor==6.7.0
django-compressor==4.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-filter==23.5
django-htmlmin==0.11.0
django-import-export==3.3.6
django-jalali==6.0.1
django-jalali-date==1.1.0
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.0
django-tailwind==3.8.0
django-watchman==1.3.0
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jdatetime==4.1.1
Jinja2==3.1.3
jsmin==3.0.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.4
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.7
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.29
pillow==10.2.0
pip-upgrader==1.4.15
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-memcached==1.62
python-slugify==8.0.3
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
rjsmin==1.2.1
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
terminaltables==3.1.10
text-unidecode==1.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
wheel==0.42.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
